#019: Added local storage
On initial API call to best buy, products get saved to local storage.
On each page refresh, local storage is checked first to see how old it
is (if more than 24 hours, clear it), and then it it is empty. If
storage is less than 24 hours old, and non empty, products are loaded
via localStorage. Else, a new API call is made.